Add logic to only remove parent if it is not uploads directory itself

+++ ils_middleware/tasks/bluecore/batch.py
@@ -47,15 +47,16 @@ def _add_other_resources(**kwargs):
         )
 
 
-def delete_upload(upload: str):
+def delete_upload(upload: str, remove_empty_parent: bool = False):
     """
     Deletes upload file
     """
     upload_path = pathlib.Path(upload)
     upload_path.unlink()
-    parent_dir = upload_path.parent
-    if parent_dir.exists() and not any(parent_dir.iterdir()):
-        parent_dir.rmdir()
+    if remove_empty_parent:
+        parent_dir = upload_path.parent
+        if parent_dir.exists() and not any(parent_dir.iterdir()):
+            parent_dir.rmdir()
 
 
 def is_zip(file_name: str) -> bool:
